Generalization of the spongeLayer class. Any number of sponges accepted
Sponges need to be specified in the ABLProperties file. As many as
needed. If no subdict is passed, the solver essentially adds zero force
and continues.

Two types are available: "Rayleigh" and "viscous". Both types have two
damping directions: "vertical" and "horizontal". For a horizontal
Rayleigh layer, "Ux" and "Uy" needs to be given.