Add option to specify the `patch` of the damping layer
Now, as opposed to specifying the start location, width, coordinate, and
step diretion, the used can simply specify `patch upper;` and the width
(either constant or time-varying) and the code will find the correct
place for the layer. Old method of specifying the sponge layers is still
valid and can used if a non-conventional layer is needed.

This commit also adds lots of comments to the code.