Modify `writeCellsCenters` to write in ascii by default
It is convenient to have the file in ascii rather than binary to be able
to quickly check it and see if the number of cells matches what the user
expects. The file is read by `wrftoof` and converted from the OpenFOAM
coordinate system to lat/lon in order to sample the wrfout files.